+++ lib/models/steps.dart
@@ -8,7 +8,7 @@ class Steps {
 
   Steps.fromJson(String date, Map<String, dynamic> json) :
       timestamp = DateFormat('yyyy-MM-dd HH:mm:ss').parse('$date ${json["time"]}'),
-      value = json["value"] ?? 0;
+      value = (json["value"] ?? 0).toDouble();
 }
 
 int getTotalStepsFromDay(List<Steps> dataSteps) {
